// src/types/generator.ts
export interface GeneratorOptions {
  upper: boolean
  lower: boolean
  number: boolean
  special: boolean
  avoidAmbiguous: boolean
  minNumbers: number
  minSpecial: number
}

export interface DisplaySettings {
  strength: boolean
  crackTime: boolean
}

export const OPTION_KEYS = {
  UPPER: 'upper',
  LOWER: 'lower',
  NUMBER: 'number',
  SPECIAL: 'special',
  AMBIGUOUS: 'avoidAmbiguous',
  MIN_NUMBERS: 'minNumbers',
  MIN_SPECIAL: 'minSpecial'
} as const

export type OnOptionChange = (_key: keyof GeneratorOptions, _value: boolean | number) => void